On supprime tout et on garde (presque) rien

+++ Code/TheGameShip/src/model/entity/Shoot.java
@@ -5,7 +5,7 @@ import java.util.UUID;
 public class Shoot extends Entity {
     private UUID ownerId;
 
-    public Shoot(UUID ownerId,String sprite, double x, double y, double radius, double hp) {
+    public Shoot(UUID ownerId,String sprite, double x, double y, double radius, int hp) {
         super(sprite,"Shoot","Shoot", x, y, radius, hp);
         this.ownerId=ownerId;
     }
